test(channelui): migrate tests for hoisted helpers into channelui
Addresses staff review's blocker #2: tests for code that lives in
channelui were accumulating in cmd/wuphf/, exercising the migrated
helpers through the lowercase alias bridge instead of the canonical
PascalCase API. The deferred test debt was about to make PR 9 (alias
removal) a 600-LoC churn PR.

Migrates the cleanly-mobile tests now — the ones that exclusively
exercise channelui-resident code:

cmd/wuphf/channelui/helpers_test.go (new):
  TestContainsString*, TestRenderTimingSummary*, TestPrettyWhenUnparsable

cmd/wuphf/channelui/messages_test.go (new):
  TestCountReplies*, TestParseTimestamp*, TestFormatShortTime*

cmd/wuphf/channelui/render_helpers_test.go (new):
  TestHumanMessageLabel*, TestRenderUnreadDivider*

cmd/wuphf/channelui/needs_you_test.go (new):
  TestBuildNeedsYouLinesPrefersBlockingRequests

Plus a tiny cmd/wuphf/channelui/test_helpers_test.go with the shared
stripANSI / joinRenderedLines helpers (the cmd/wuphf copies stay until
the alias bridge comes down).

Tests for code that still lives in cmd/wuphf (buildSkillLines,
buildOneOnOneMessageLines, requestCalendarEvents,
calendarParticipantsForRequest, summarizeUnreadMessages,
defaultHumanMessageTitle, sliceRenderedLines, buildRequestLines,
buildPolicyLines) intentionally stay put — they'll move with their
production code in subsequent PRs.
