+++ cmd/wuphf/channel_needs_you_test.go
@@ -5,26 +5,6 @@ import (
 	"testing"
 )
 
-func TestBuildNeedsYouLinesPrefersBlockingRequests(t *testing.T) {
-	requests := []channelInterview{
-		{ID: "req-1", Kind: "approval", Status: "pending", Title: "Optional note", Question: "Optional note?", From: "pm"},
-		{ID: "req-2", Kind: "approval", Status: "pending", Title: "Ship launch copy", Question: "Ship launch copy?", Context: "Need approval before publishing.", From: "ceo", Blocking: true, RecommendedID: "approve"},
-	}
-
-	lines := buildNeedsYouLines(requests, 96)
-	plain := stripANSI(joinRenderedLines(lines))
-
-	if !strings.Contains(plain, "Needs attention") {
-		t.Fatalf("expected needs-attention separator, got %q", plain)
-	}
-	if !strings.Contains(plain, "Ship launch copy") {
-		t.Fatalf("expected blocking request title, got %q", plain)
-	}
-	if !strings.Contains(plain, "The team is paused until you answer.") {
-		t.Fatalf("expected blocking request guidance, got %q", plain)
-	}
-}
-
 func TestCurrentMainViewportLinesPrependsNeedsYouStrip(t *testing.T) {
 	m := newChannelModel(false)
 	m.width = 120
